backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
boto3==1.40.0
pydantic-settings==2.10.1
